+++ config/application.rb
@@ -33,6 +33,7 @@ class Application < Rails::Application
   config.eager_load_paths << "#{Rails.root}/app/models/components"
   config.eager_load_paths << "#{Rails.root}/app/controllers/components"
   config.eager_load_paths << "#{Rails.root}/app/services"
+  config.eager_load_paths << "#{Rails.root}/app/services/concerns"
   config.eager_load_paths << "#{Rails.root}/app/notifiers"
 
   config.x.public_download_folder = 'downloads'
